Refactor foodsController out of server.js
* Relocate 2nd-argument functions from server requests to controller
* Only access models from controllers
